fix: harden input handling, mock wiring, and config payload (v0.0.36)
Code review found several concrete defects. This release addresses them.

Security / hardening:
- get_calibration_history(): percent-encode the `sensor` argument before
  concatenating it into the request URL. A value with spaces, `#`, `&`
  or other reserved chars previously produced a malformed URL or injected
  an extra query parameter.
- get_log(): validate `log_type` against the documented allow-list
  (actions/switching/onewire) and raise VioletPoolAPIError otherwise.
- restore_calibration(): sanitize `sensor` and `timestamp` (length bounds,
  character filtering) before posting, matching the treatment of every
  key in set_config().

Fixes:
- _sanitize_config_payload(): preserve `int` and `bool` values instead of
  routing them through sanitize_numeric(), which always returns float.
  Integer flags (e.g. DOS_*_use from set_dosage_enabled()) were being sent
  as `1.0`/`0.0`. bool is now serialized as 0/1.
- set_config(): drop the undocumented retryable=True override so it behaves
  like every other state-changing POST (non-retryable on 5xx).
- mock server: dispatch OMNI queries from handle_set_function_manually to
  handle_set_omni_position. The handler existed but was never routed, so
  set_omni_position() returned `OK\nOMNI\n...` (wrong) and never updated
  omni_position in /mock/state. Now returns `OK\nOMNITRONIC\n...`.

Cleanup:
- Remove duplicate InputSanitizer.validate_duration/validate_speed that
  collided by name with the raising validators in _api_model but silently
  clamped values. Unused dead code.
- Type `duration`/`last_value` as int | None (was float | None) to match
  the actual whole-number validation/truncation behaviour.
- Fix mojibake in docstrings/comments and drop redundant
  except (Exception): parentheses.

Tests: +10 regression tests covering each fix; smoke test gains OMNI
end-to-end coverage. 144 unit + 68 smoke tests pass; ruff & mypy clean.

Bumps version to 0.0.36.

+++ violet_poolcontroller_api/_api_mixin.py
@@ -49,8 +49,8 @@ class APIClientMixin:
         key: str,
         action: str,
         *,
-        duration: float | None = None,
-        last_value: float | None = None,
+        duration: int | None = None,
+        last_value: int | None = None,
     ) -> dict[str, Any]:
         raise NotImplementedError
 
@@ -59,7 +59,7 @@ class APIClientMixin:
         key: str,
         action: str,
         *,
-        duration: float | None = None,
+        duration: int | None = None,
     ) -> dict[str, Any]:
         raise NotImplementedError
 
@@ -68,8 +68,8 @@ class APIClientMixin:
         key: str,
         action: str,
         *,
-        duration: float | None = None,
-        last_value: float | None = None,
+        duration: int | None = None,
+        last_value: int | None = None,
     ) -> str:
         raise NotImplementedError
